+++ CHANGELOG.rst
@@ -2,6 +2,15 @@
 Changelog
 =========
 
+0.9.4
+=====
+
+* Refactor forms app into independent project
+* Deprecation warning for forms app
+* Add dark mode compatibility with django CMS 3.11
+* Remove strong dependency on djangocms-icon
+* Sync github and pypi releases
+
 0.9.1
 =====